NU-1080 update after review
