--- designer/client/src/components/graph/joint-events.ts
@@ -1,22 +1,0 @@
-export enum Events {
-    LINK_CONNECT = "link:connect",
-    LINK_DISCONNECT = "link:disconnect",
-    LINK_MOUSEOVER = "link:mouseover",
-    LINK_MOUSEENTER = "link:mouseenter",
-    LINK_MOUSELEAVE = "link:mouseleave",
-    CELL_POINTERUP = "cell:pointerup",
-    CELL_POINTERDOWN = "cell:pointerdown",
-    CELL_POINTERCLICK = "cell:pointerclick",
-    CELL_POINTERDBLCLICK = "cell:pointerdblclick",
-    CELL_MOUSEOVER = "cell:mouseover",
-    CELL_MOUSEOUT = "cell:mouseout",
-    CELL_MOVED = "cellCustom:moved",
-    BLANK_POINTERDOWN = "blank:pointerdown",
-    BLANK_POINTERUP = "blank:pointerup",
-    BLANK_POINTERMOVE = "blank:pointermove",
-    BLANK_MOUSEOVER = "blank:mouseover",
-    REMOVE = "remove",
-    ADD = "add",
-    CHANGE_POSITION = "change:position",
-    CHANGE_DRAG_OVER = "change:draggedOver",
-}
